search/requirements.txt
flask
requests
chromadb
elasticsearch
psycopg2-binary
openai